Clean up data/ directory
